% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/article_ezANOVA.R
\name{article.ezANOVA}
\alias{article.ezANOVA}
\title{Article-output ezANOVA}
\usage{
article.ezANOVA(aov = NULL)
}
\arguments{
\item{aov}{the output object of an ezANOVA}
}
\value{
string vector giving the formatted article text
article.ezANOVA()
}
\description{
Produces article-ready APA output for an ezANOVA object
}
\keyword{APA}
\keyword{article}
\keyword{ezANOVA}
