% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/article_ttest.R
\name{article.ttest}
\alias{article.ttest}
\title{Article-output for a t.test (Simple)}
\usage{
article.ttest(obj = NULL, unit = "ms")
}
\arguments{
\item{obj}{the output object of an t.test function (stats package)}

\item{unit}{the measurement unit, defaults to 'ms'}
}
\value{
string vector giving the formatted article text
article.ttest()
}
\description{
Produces article-ready APA output for an t.test-object, including test statistic, p-value and confidence intervals
}
\keyword{APA}
\keyword{article}
\keyword{t.test}
