% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/article_ttest.R
\name{article.ttest_full}
\alias{article.ttest_full}
\title{Comprehensive Article-output for a t.test}
\usage{
article.ttest_full(obj = NULL, dz = NULL, BF = NULL, unit = "ms")
}
\arguments{
\item{obj}{the output object of an t.test function (stats package)}

\item{dz}{output from cohensD() function (CohensD package)}

\item{BF}{output from ttestBF()-function (BayesFactor package)}

\item{unit}{the measurement unit, defaults to 'ms'}
}
\value{
string vector giving the formatted article text
article.ttest_full()
}
\description{
Produces article-ready APA output for an t.test-object, including test statistic, p-value, Cohen's D, BayesFactor and Highest Posterior Density Interval
}
\keyword{APA}
\keyword{article}
\keyword{t.test}
